Trie class refactor for clarity In preparation for future performance improvements
