+++ rskj-core/src/main/java/co/rsk/trie/IterationElement.java
@@ -1,0 +1,46 @@
+/*
+ * This file is part of RskJ
+ * Copyright (C) 2022 RSK Labs Ltd.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package co.rsk.trie;
+
+public class IterationElement {
+
+        private final TrieKeySlice nodeKey;
+        private final Trie node;
+
+        public IterationElement(final TrieKeySlice nodeKey, final Trie node) {
+            this.nodeKey = nodeKey;
+            this.node = node;
+        }
+
+        public Trie getNode() {
+            return node;
+        }
+
+        public final TrieKeySlice getNodeKey() {
+            return nodeKey;
+        }
+
+        public String toString() {
+            byte[] encodedFullKey = nodeKey.expand();
+            StringBuilder ouput = new StringBuilder();
+            for (byte b : encodedFullKey) {
+                ouput.append( b == 0 ? '0': '1');
+            }
+            return ouput.toString();
+        }
+}
