+++ rskj-core/src/main/java/co/rsk/trie/Trie.java
@@ -739,7 +739,7 @@ public class Trie {
         encoded = buffer.array();
     }
 
-    private Trie retrieveNode(byte implicitByte) {
+    protected Trie retrieveNode(byte implicitByte) {
         return getNodeReference(implicitByte).getNode().orElse(null);
     }
 
@@ -1147,287 +1147,6 @@ public class Trie {
 
         message.get(bytes);
         return new VarInt(bytes, 0);
-    }
-
-    /**
-     * Returns the leftmost node that has not yet been visited that node is normally on top of the stack
-     */
-    private static class InOrderIterator implements Iterator<IterationElement> {
-
-        private final Deque<IterationElement> visiting;
-
-        public InOrderIterator(Trie root) {
-            Objects.requireNonNull(root);
-            TrieKeySlice traversedPath = root.getSharedPath();
-            this.visiting = new LinkedList<>();
-            // find the leftmost node, pushing all the intermediate nodes onto the visiting stack
-            visiting.push(new IterationElement(traversedPath, root));
-            pushLeftmostNode(traversedPath, root);
-            // now the leftmost unvisited node is on top of the visiting stack
-        }
-
-        /**
-         * return the leftmost node that has not yet been visited that node is normally on top of the stack
-         */
-        @Override
-        @SuppressWarnings("squid:S2272") // NoSuchElementException is thrown by {@link Deque#pop()} when it's empty
-        public IterationElement next() {
-            IterationElement visitingElement = visiting.pop();
-            Trie node = visitingElement.getNode();
-            // if the node has a right child, its leftmost node is next
-            Trie rightNode = node.retrieveNode((byte) 0x01);
-            if (rightNode != null) {
-                TrieKeySlice rightNodeKey = visitingElement.getNodeKey().rebuildSharedPath((byte) 0x01, rightNode.getSharedPath());
-                visiting.push(new IterationElement(rightNodeKey, rightNode)); // push the right node
-                // find the leftmost node of the right child
-                pushLeftmostNode(rightNodeKey, rightNode);
-                // note "node" has been replaced on the stack by its right child
-            } // else: no right subtree, go back up the stack
-            // next node on stack will be next returned
-            return visitingElement;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !visiting.isEmpty(); // no next node left
-        }
-
-        /**
-         * Find the leftmost node from this root, pushing all the intermediate nodes onto the visiting stack
-         *
-         * @param nodeKey
-         * @param node the root of the subtree for which we are trying to reach the leftmost node
-         */
-        private void pushLeftmostNode(TrieKeySlice nodeKey, Trie node) {
-            // find the leftmost node
-            Trie leftNode = node.retrieveNode((byte) 0x00);
-            if (leftNode != null) {
-                TrieKeySlice leftNodeKey = nodeKey.rebuildSharedPath((byte) 0x00, leftNode.getSharedPath());
-                visiting.push(new IterationElement(leftNodeKey, leftNode)); // push the left node
-                pushLeftmostNode(leftNodeKey, leftNode); // recurse on next left node
-            }
-        }
-    }
-
-    private class PreOrderIterator implements Iterator<IterationElement> {
-
-        private final Deque<IterationElement> visiting;
-        private boolean stopAtAccountDepth;
-
-        public PreOrderIterator(Trie root,boolean stopAtAccountDepth) {
-            Objects.requireNonNull(root);
-            this.stopAtAccountDepth = stopAtAccountDepth;
-            TrieKeySlice traversedPath = root.getSharedPath();
-            this.visiting = new LinkedList<>();
-            this.visiting.push(new IterationElement(traversedPath, root));
-        }
-
-        @Override
-        @SuppressWarnings("squid:S2272") // NoSuchElementException is thrown by {@link Deque#pop()} when it's empty
-        public IterationElement next() {
-            IterationElement visitingElement = visiting.pop();
-            Trie node = visitingElement.getNode();
-            TrieKeySlice nodeKey = visitingElement.getNodeKey();
-
-            int nodeKeyLength = nodeKey.length();
-            // If we're stoping at accounts, do not add children
-            if ((!stopAtAccountDepth) || (nodeKeyLength < (1 + TrieKeyMapper.SECURE_KEY_SIZE + RskAddress.LENGTH_IN_BYTES) * Byte.SIZE))
-            {
-                // need to visit the left subtree first, then the right since a stack is a LIFO, push the right subtree first,
-                // then the left
-                Trie rightNode = node.retrieveNode((byte) 0x01);
-                if (rightNode != null) {
-                    TrieKeySlice rightNodeKey = nodeKey.rebuildSharedPath((byte) 0x01, rightNode.getSharedPath());
-                    visiting.push(new IterationElement(rightNodeKey, rightNode));
-                }
-                Trie leftNode = node.retrieveNode((byte) 0x00);
-                if (leftNode != null) {
-                    TrieKeySlice leftNodeKey = nodeKey.rebuildSharedPath((byte) 0x00, leftNode.getSharedPath());
-                    visiting.push(new IterationElement(leftNodeKey, leftNode));
-                }
-            }
-            // may not have pushed anything.  If so, we are at the end
-            return visitingElement;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !visiting.isEmpty(); // no next node left
-        }
-    }
-
-    private class PostOrderIterator implements Iterator<IterationElement> {
-
-        private final Deque<IterationElement> visiting;
-        private final Deque<Boolean> visitingRightChild;
-
-        public PostOrderIterator(Trie root) {
-            Objects.requireNonNull(root);
-            TrieKeySlice traversedPath = root.getSharedPath();
-            this.visiting = new LinkedList<>();
-            this.visitingRightChild = new LinkedList<>();
-            // find the leftmost node, pushing all the intermediate nodes onto the visiting stack
-            visiting.push(new IterationElement(traversedPath, root));
-            visitingRightChild.push(Boolean.FALSE);
-            pushLeftmostNodeRecord(traversedPath, root);
-            // the node on top of the visiting stack is the next one to be visited, unless it has a right subtree
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !visiting.isEmpty(); // no next node left
-        }
-
-        @Override
-        @SuppressWarnings("squid:S2272") // NoSuchElementException is thrown by {@link Deque#element()} when it's empty
-        public IterationElement next() {
-            IterationElement visitingElement = visiting.element();
-            Trie node = visitingElement.getNode();
-            Trie rightNode = node.retrieveNode((byte) 0x01);
-            if (rightNode == null || visitingRightChild.peek()) { // no right subtree, or right subtree already visited
-                // already visited right child, time to visit the node on top
-                visiting.removeFirst(); // it was already picked
-                visitingRightChild.removeFirst(); // it was already picked
-                return visitingElement;
-            } else { // now visit this node's right subtree
-                // mark that we're visiting this element's right subtree
-                visitingRightChild.removeFirst();
-                visitingRightChild.push(Boolean.TRUE);
-
-                TrieKeySlice rightNodeKey = visitingElement.getNodeKey().rebuildSharedPath((byte) 0x01, rightNode.getSharedPath());
-                visiting.push(new IterationElement(rightNodeKey, rightNode)); // push the right node
-                visitingRightChild.push(Boolean.FALSE); // we're visiting the left subtree of the right node
-
-                // now push everything down to the leftmost node in the right subtree
-                pushLeftmostNodeRecord(rightNodeKey, rightNode);
-                return next(); // use recursive call to visit that node
-            }
-        }
-
-        /**
-         * Find the leftmost node from this root, pushing all the intermediate nodes onto the visiting stack
-         * and also stating that each is a left child of its parent
-         * @param nodeKey
-         * @param node the root of the subtree for which we are trying to reach the leftmost node
-         */
-        private void pushLeftmostNodeRecord(TrieKeySlice nodeKey, Trie node) {
-            // find the leftmost node
-            Trie leftNode = node.retrieveNode((byte) 0x00);
-            if (leftNode != null) {
-                TrieKeySlice leftNodeKey = nodeKey.rebuildSharedPath((byte) 0x00, leftNode.getSharedPath());
-                visiting.push(new IterationElement(leftNodeKey, leftNode)); // push the left node
-                visitingRightChild.push(Boolean.FALSE); // record that it is on the left
-                pushLeftmostNodeRecord(leftNodeKey, leftNode); // continue looping
-            }
-        }
-    }
-
-    public static class IterationElement {
-        private final TrieKeySlice nodeKey;
-        private final Trie node;
-
-        public IterationElement(final TrieKeySlice nodeKey, final Trie node) {
-            this.nodeKey = nodeKey;
-            this.node = node;
-        }
-
-        public Trie getNode() {
-            return node;
-        }
-
-        public final TrieKeySlice getNodeKey() {
-            return nodeKey;
-        }
-
-        public String toString() {
-            byte[] encodedFullKey = nodeKey.encode();
-            StringBuilder ouput = new StringBuilder();
-            for (byte b : encodedFullKey) {
-                ouput.append( b == 0 ? '0': '1');
-            }
-            return ouput.toString();
-        }
-    }
-
-    private static class SharedPathSerializer {
-        private final TrieKeySlice sharedPath;
-        private final int lshared;
-
-        private SharedPathSerializer(TrieKeySlice sharedPath) {
-            this.sharedPath = sharedPath;
-            this.lshared = this.sharedPath.length();
-        }
-
-        private int serializedLength() {
-            if (!isPresent()) {
-                return 0;
-            }
-
-            return lsharedSize() + PathEncoder.calculateEncodedLength(lshared);
-        }
-
-        public boolean isPresent() {
-            return lshared > 0;
-        }
-
-        private void serializeInto(ByteBuffer buffer) {
-            if (!isPresent()) {
-                return;
-            }
-
-            if (1 <= lshared && lshared <= 32) {
-                // first byte in [0..31]
-                buffer.put((byte) (lshared - 1));
-            } else if (160 <= lshared && lshared <= 382) {
-                // first byte in [32..254]
-                buffer.put((byte) (lshared - 128));
-            } else {
-                buffer.put((byte) 255);
-                buffer.put(new VarInt(lshared).encode());
-            }
-
-            buffer.put(this.sharedPath.encode());
-        }
-
-        private int lsharedSize() {
-            if (!isPresent()) {
-                return 0;
-            }
-
-            if (1 <= lshared && lshared <= 32) {
-                return 1;
-            }
-
-            if (160 <= lshared && lshared <= 382) {
-                return 1;
-            }
-
-            return 1 + VarInt.sizeOf(lshared);
-        }
-
-        private static TrieKeySlice deserialize(ByteBuffer message, boolean sharedPrefixPresent) {
-            if (!sharedPrefixPresent) {
-                return TrieKeySlice.empty();
-            }
-
-            int lshared;
-            // upgrade to int so we can compare positive values
-            int lsharedFirstByte = Byte.toUnsignedInt(message.get());
-            if (0 <= lsharedFirstByte && lsharedFirstByte <= 31) {
-                // lshared in [1..32]
-                lshared = lsharedFirstByte + 1;
-            } else if (32 <= lsharedFirstByte && lsharedFirstByte <= 254) {
-                // lshared in [160..382]
-                lshared = lsharedFirstByte + 128;
-            } else {
-                lshared = (int) readVarInt(message).value;
-            }
-
-            int lencoded = PathEncoder.calculateEncodedLength(lshared);
-            byte[] encodedKey = new byte[lencoded];
-            message.get(encodedKey);
-            return TrieKeySlice.fromEncoded(encodedKey, 0, lshared, lencoded);
-        }
     }
 
     // Additional auxiliary methods for Merkle Proof
